requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==2.0.1
openai==1.40.0
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.11.0
urllib3==2.2.2
Werkzeug==3.0.3
flask_login==0.6.3
kinde-python-sdk==1.2.6
flask_limiter==3.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
scenedetect==0.6.4
opencv-python==4.10.0.84